Exit Confirmation Page Added
When pressing the Close Button on the Upper Right Corner or when pressing the Exit Application in the Menu Bar, a message will pop up confirming the application exit. The pop up also asks if the user wants to export the current Inventory to a CSV file (Code is still Unavailable).

Data can be considered safe, especially that the confirmation page blocks any form of accidental closure to the program.

++ Modified and Made few Adjustments to the Menu Bar